ponderada4/requirements.txt
streamlit==1.20.0
openai==0.27.2
python-dotenv